Update README.md
Hi, I just tried `faker_airtravel` and realised that the readme is not up-to-date. In the given readme examples some of the keys in the returned dictionary of `fake.airport_object()` and `fake.flight()` are capitalized. In reality, this is not the case. Therefore I created this PR to update the readme.